refactor/#160: 기존의 로그인 리액터 이동 및 컨벤션 적용
